[kernel] print some multiboot info, and fix terminal_writeuint32

+++ kernel/terminal.h
@@ -43,6 +43,6 @@ void terminal_putentryat(char c, uint8_t color, size_t x, size_t y);
 void terminal_newline();
 void terminal_putchar(char c);
 void terminal_writestring(const char* data);
-int terminal_writeuint32(uint32_t integer, uint32_t base);
+int terminal_writeuint32(uint32_t integer, uint8_t base);
 
 #endif
